documentation

+++ src/Team.cpp
@@ -1,5 +1,5 @@
 #include "./Team.hpp"
-Team::Team(string name, string comuna, string estadio, double lat, double lng) : coordinate(lat, lng)
+Team::Team(string name, string comuna, string estadio, long double lat, long double lng) : coordinate(lat, lng)
 {
     this->name = name;
     this->comuna = comuna;
